refactor: split engine ops bridge modules (#346)
